WIP Add filters flyout
