+++ demo/app/src/components/Sidebar/Sidebar.js
@@ -2,16 +2,14 @@ import RingLoader from 'react-spinners/RingLoader';
 import Button from '@components/Button';
 import {
   HorizontalSlidersLines,
-  AlignJustifyDown,
-  ArrowRightLine
+  AlignJustifyDown
 } from '@vectopus/atlas-icons-react';
 import styles from "./Sidebar.module.scss";
 
 import { useAppContext } from 'src/context/appContext';
 import OpportunityList from './OpportunityList';
-import { useEffect, useState } from 'react';
-import { ALL_PROVIDERS as providers } from 'src/utils/constants';
 import OpportunityDetail from './OpportunityDetail';
+import RefinePane from './RefinePane';
 
 export default function Sidebar(props) {
   const {
@@ -19,54 +17,8 @@ export default function Sidebar(props) {
     isLoadingOpps,
     errorOpps,
     openFilters,
-    setOpenFilters,
-    products,
-    userParams,
-    setUserParams
+    setOpenFilters
   } = useAppContext();
-
-  function providerSelectHandler(e) {
-    setUserParams({
-      ...userParams,
-      provider: e.target.value,
-      product: 'all'
-    })
-  }
-
-  function productSelectHandler(e) {
-    setUserParams({
-      ...userParams,
-      product: e.target.value
-    })
-  }
-
-  function isProviderSelected(provider) {
-    return provider === userParams.provider;
-  }
-
-  function isProductSelected(product) {
-    return product === userParams.product;
-  }
-
-  function getProductOptions(products){
-    return products && products.filter(product => product).map(({ title, id }) => {
-      return (
-        <option
-          key={id}
-          value={id}
-          selected={isProductSelected(id)}
-        >
-          {title}
-        </option>
-      );
-    });
-  }
-
-  function displayProducts() {
-    return userParams.provider !== 'all' ?
-      getProductOptions(products[userParams.provider])
-      : getProductOptions(Object.values(products).flat())
-  }
 
   const filterButtonClass = openFilters
     ? styles.filterButtonOpen
@@ -94,55 +46,7 @@ export default function Sidebar(props) {
 
           {!!errorOpps && <div>There was error</div>}
 
-          {!!openFilters && (
-            <div className={styles.filtersFlyout}>
-              <div className={styles.filtersTopBar}>
-                <h4>Filters</h4>
-                <Button
-                  className={styles.closeButton}
-                  onClick={() => setOpenFilters(false)}
-                >
-                  <ArrowRightLine size={12} />
-                </Button>
-              </div>
-              <div className={styles.filtersBody}>
-                <div>
-                  <span>Providers: </span>
-                  <select onChange={providerSelectHandler}>
-                    <option
-                        value='all'
-                        key='all'
-                        selected
-                      >
-                        All
-                      </option>
-                    {providers.map((provider) => (
-                      <option
-                        value={provider.id}
-                        key={provider.id}
-                        selected={isProviderSelected(provider.id)}
-                      >
-                        {provider.name}
-                      </option>
-                    ))}
-                  </select>
-                </div>
-                <div>
-                  <div className={styles.blockLabel}>Products: </div>
-                  <select onChange={productSelectHandler}>
-                      <option
-                        value='all'
-                        key='all'
-                        selected
-                      >
-                        All
-                      </option>
-                      {displayProducts()}
-                    </select>
-                </div>
-              </div>
-            </div>
-          )}
+          {!!openFilters && <RefinePane />}
         </>
       )}
       {!!isLoadingOpps && !errorOpps && (
