+++ demo/app/src/components/Sidebar/Sidebar.module.scss
@@ -7,7 +7,7 @@
   left: 0;
   width: 350px;
   height: 100%;
-  z-index: 1000;
+  z-index: 1001;
   background-color: white;
 }
 
@@ -129,13 +129,34 @@
   }
 
   .filtersBody {
-    padding-left: 5px;
     text-align: left;
 
     .blockLabel {
       display: block;
-      text-align: left;
       padding: 0;
+    }
+
+    .refineLabel {
+      margin-top: 10px;
+      padding: 5px;
+      display: block;
+      background-color: lightgrey;
+      width: 100%;
+      text-align: left;
+    }
+
+    .constraintLabel {
+      background-color: none;
+      padding-left: 10px;
+    }
+
+    .refineSelection {
+      display: block;
+      color: grey;
+      width: 80%;
+      margin-top: 10px;
+      margin-left: auto;
+      margin-right: auto;
     }
   }
 }
@@ -158,7 +179,7 @@
   border: 3px solid #1C5F78;
   border-right: none;
   box-shadow: 5px 0 0 0 white;
-  z-index: 1001;
+  z-index: 1002;
 }
 
 .opportunityPreview > h1 {
